+++ lib/teamlab/modules/projects/projects_comments.rb
@@ -30,5 +30,21 @@ module Teamlab
     def delete_comment(comment_id)
       @request.delete(['comment', comment_id.to_s])
     end
+
+    # Adds a project comment with the parameters specified in the request
+    # @param entity_id [String] Entity ID
+    # @param content [String] Comment text
+    # @param type [String, Symbol] Comment type (message or task)
+    # @return [Hash] Comment information
+    def add_project_comment(entity_id, content, type, options = {})
+      @request.post(%w[comment], { entityid: entity_id, content: content, type: type.to_s }.merge(options))
+    end
+
+    # Get a preview of a project comment with the ID specified in the request
+    # @param htmltext [String] Comment text in the HTML format
+    # @return [Hash] Comment information
+    def get_a_comment_preview(htmltext, options = {})
+      @request.post(%w[comment preview], { htmltext: htmltext }.merge(options))
+    end
   end
 end