+++ CHANGELOG.md
@@ -21,6 +21,8 @@
 * Add `People#start_remove` method
 * Add `People#remove_progress` method
 * Add `People#terminate_remove` method
+* Add `Projects#get_a_comment_preview` method
+* Add `Projects#add_project_comment` method
 
 ### Refactor